Move defaultProps to compat
