+++ compat/test/browser/component.test.js
@@ -76,6 +76,91 @@ describe('components', () => {
 		});
 	});
 
+	describe('defaultProps', () => {
+		it('should apply default props on initial render', () => {
+			class WithDefaultProps extends React.Component {
+				constructor(props, context) {
+					super(props, context);
+					expect(props).to.be.deep.equal({
+						fieldA: 1,
+						fieldB: 2,
+						fieldC: 1,
+						fieldD: 2
+					});
+				}
+				render() {
+					return <div />;
+				}
+			}
+			WithDefaultProps.defaultProps = { fieldC: 1, fieldD: 1 };
+			React.render(
+				<WithDefaultProps fieldA={1} fieldB={2} fieldD={2} />,
+				scratch
+			);
+		});
+
+		it('should apply default props on rerender', () => {
+			let doRender;
+			class Outer extends React.Component {
+				constructor() {
+					super();
+					this.state = { i: 1 };
+				}
+				componentDidMount() {
+					doRender = () => this.setState({ i: 2 });
+				}
+				render(props, { i }) {
+					return <WithDefaultProps fieldA={1} fieldB={i} fieldD={i} />;
+				}
+			}
+			class WithDefaultProps extends React.Component {
+				constructor(props, context) {
+					super(props, context);
+					this.ctor(props, context);
+				}
+				ctor() {}
+				componentWillReceiveProps() {}
+				render() {
+					return <div />;
+				}
+			}
+			WithDefaultProps.defaultProps = { fieldC: 1, fieldD: 1 };
+
+			let proto = WithDefaultProps.prototype;
+			sinon.spy(proto, 'ctor');
+			sinon.spy(proto, 'componentWillReceiveProps');
+			sinon.spy(proto, 'render');
+
+			React.render(<Outer />, scratch);
+			doRender();
+
+			const PROPS1 = {
+				fieldA: 1,
+				fieldB: 1,
+				fieldC: 1,
+				fieldD: 1
+			};
+
+			const PROPS2 = {
+				fieldA: 1,
+				fieldB: 2,
+				fieldC: 1,
+				fieldD: 2
+			};
+
+			expect(proto.ctor).to.have.been.calledWithMatch(PROPS1);
+			expect(proto.render).to.have.been.calledWithMatch(PROPS1);
+
+			rerender();
+
+			// expect(proto.ctor).to.have.been.calledWith(PROPS2);
+			expect(proto.componentWillReceiveProps).to.have.been.calledWithMatch(
+				PROPS2
+			);
+			expect(proto.render).to.have.been.calledWithMatch(PROPS2);
+		});
+	});
+
 	describe('UNSAFE_* lifecycle methods', () => {
 		it('should support UNSAFE_componentWillMount', () => {
 			let spy = sinon.spy();
