+++ src/create-element.js
@@ -31,16 +31,6 @@ export function createElement(type, props, children) {
 		normalizedProps.children = children;
 	}
 
-	// If a Component VNode, check for and apply defaultProps
-	// Note: type may be undefined in development, must never error here.
-	if (typeof type == 'function' && type.defaultProps != null) {
-		for (i in type.defaultProps) {
-			if (normalizedProps[i] === undefined) {
-				normalizedProps[i] = type.defaultProps[i];
-			}
-		}
-	}
-
 	return createVNode(type, normalizedProps, key, ref, null);
 }
